rag-llm/requirements.txt
flask
langchain
langchain-openai
langchain-groq
langchain-google-genai
python-dotenv
pydantic
pymongo